pyproject.toml
[project]
name = "nse-xbrl-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arelle-release>=2.38.17",
    "bs4>=0.0.2",
    "httpx[http2]>=0.28.1",
    "lxml>=5.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
]
